COMMERCE-10086 Minor Refactor SF. Move to commerce/order
