+++ modules/apps/commerce/commerce-service/src/main/java/com/liferay/commerce/internal/model/listener/CommerceOrderItemModelListener.java
@@ -18,7 +18,7 @@ import com.liferay.commerce.constants.CommerceOrderConstants;
 import com.liferay.commerce.model.CommerceOrder;
 import com.liferay.commerce.model.CommerceOrderItem;
 import com.liferay.commerce.order.engine.CommerceOrderEngine;
-import com.liferay.commerce.util.CommerceOrderThreadLocal;
+import com.liferay.commerce.order.CommerceOrderThreadLocal;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
